// src/context/StoreProvider/types/createMeta.ts
export interface CreateMeta {
  expand: string;
  projects: ProjectElement[];
}

export interface ProjectElement {
  expand: string;
  self: string;
  id: string;
  key: AllowedValueKey;
  name: string;
  avatarUrls: { [key: string]: string };
  issuetypes: Issuetype[];
}

export interface Issuetype {
  self: string;
  id: string;
  description: string;
  iconUrl: string;
  name: string;
  untranslatedName: string;
  subtask: boolean;
  expand: Expand;
  fields: Record<string, Assignee | Attachment>;
  scope?: Scope;
}
// ...
export enum AllowedValueKey {
  CSM = "CSM",
  Demo = "DEMO",
  Tp = "TP",
}